+++ app/src/main/java/com/company/mobile/android/appname/app/bufferoos/master/BufferoosNavigationCommand.kt
@@ -1,0 +1,6 @@
+package com.company.mobile.android.appname.app.bufferoos.master
+
+sealed class BufferoosNavigationCommand {
+
+    object GoToDetailsView : BufferoosNavigationCommand()
+}