+++ app/src/main/java/com/company/mobile/android/appname/app/bufferoos/viewmodel/BufferoosState.kt
@@ -15,7 +15,7 @@ sealed class BufferoosState(
     )
 
     data class Error(private val message: String? = null) : BufferoosState(
-        ResourceState.SUCCESS,
+        ResourceState.ERROR,
         errorMessage = message
     )