+++ app/src/main/java/com/company/mobile/android/appname/app/bufferoos/viewmodel/BufferoosViewModel.kt
@@ -3,16 +3,23 @@ package com.company.mobile.android.appname.app.bufferoos.viewmodel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import com.company.mobile.android.appname.app.bufferoos.master.BufferoosNavigationCommand
+import com.company.mobile.android.appname.app.bufferoos.master.BufferoosNavigationCommand.GoToDetailsView
 import com.company.mobile.android.appname.app.bufferoos.viewmodel.BufferoosState.Error
 import com.company.mobile.android.appname.app.bufferoos.viewmodel.BufferoosState.Loading
 import com.company.mobile.android.appname.app.bufferoos.viewmodel.BufferoosState.Success
+import com.company.mobile.android.appname.app.common.viewmodel.SingleLiveEvent
 import com.company.mobile.android.appname.domain.bufferoo.interactor.GetBufferoos
+import com.company.mobile.android.appname.model.bufferoo.Bufferoo
 import io.reactivex.disposables.Disposable
 
 class BufferoosViewModel(val getBufferoos: GetBufferoos) : ViewModel() {
 
     private val bufferoosLiveData: MutableLiveData<BufferoosState> = MutableLiveData()
+    private val selectedBufferooLiveData: MutableLiveData<Bufferoo> = MutableLiveData()
+    private var bufferoos: List<Bufferoo> = emptyList()
     private var disposable: Disposable? = null
+    val bufferoosNavigationEvent = SingleLiveEvent<BufferoosNavigationCommand>()
 
     override fun onCleared() {
         disposable?.dispose()
@@ -23,11 +30,23 @@ class BufferoosViewModel(val getBufferoos: GetBufferoos) : ViewModel() {
         return bufferoosLiveData
     }
 
+    fun getSelectedBufferoo(): LiveData<Bufferoo> {
+        return selectedBufferooLiveData
+    }
+
+    fun select(position: Int) {
+        // Store selected item for details view
+        selectedBufferooLiveData.value = bufferoos[position]
+        // Go to details view
+        bufferoosNavigationEvent.value = GoToDetailsView
+    }
+
     fun fetchBufferoos() {
         bufferoosLiveData.postValue(Loading)
         disposable = getBufferoos.execute()
             .subscribe({
-                bufferoosLiveData.postValue(Success(it))
+                bufferoos = it
+                bufferoosLiveData.postValue(Success(bufferoos))
             }, {
                 bufferoosLiveData.postValue(Error(it.message))
             })